Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/matchMedia-expected.txt


FAIL window.matchMedia exists assert_inherits: property "matchMedia" found on object expected in prototype chain
PASS window.matchMedia is a Function 
FAIL window.matchMedia("all") Can't find variable: MediaQueryList
PASS MediaQueryList.media exists 
PASS MediaQueryList.media is readonly 
PASS MediaQueryList.media for "all" 
PASS MediaQueryList.matches exists 
PASS MediaQueryList.matches is readonly 
PASS MediaQueryList.matches for "all" 
PASS MediaQueryList.addListener exists 
PASS MediaQueryList.addListener is a Function 
PASS MediaQueryList.removeListener exists 
PASS MediaQueryList.removeListener is a Function 
FAIL MediaQueryList.media syntax error Can't find variable: MediaQueryList
PASS MediaQueryList.matches for "not all" 
PASS MediaQueryList.matches for "(max-width: 199px), all and (min-width: 200px)" 
PASS MediaQueryList.matches for "(min-aspect-ratio: 1/1)" 
PASS MediaQueryList.matches for "(width: 200px)" 
PASS MediaQueryList.matches for "(max-height: 50px)" 
PASS MediaQueryList.matches for "(min-width: 150px)" 
PASS Resize iframe from 200x100 to 200x50, then to 100x50 
PASS Listeners are called in the order which they have been added 
PASS Listener added twice is only called once. 
